Rename human_task_watcher module
